Backend clean up and restructure
