+++ backend/core/fileupload/serializers.py
@@ -4,6 +4,9 @@ from core.fileupload.models import File
 
 
 class FilesSerializer(serializers.HyperlinkedModelSerializer):
+    """
+    A serializer for defining which file attributes should be converted to JSON
+    """
     owner = serializers.ReadOnlyField(source='owner.username')
 
     class Meta:
