+++ backend/core/user/forms.py
@@ -1,5 +1,4 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm as UserFrom
 from django.core.exceptions import ValidationError
 from django.contrib.auth.forms import ReadOnlyPasswordHashField
 from .models import User
@@ -7,7 +6,7 @@ from .models import User
 
 class AdminUserCreationForm(forms.ModelForm):
     """
-    A form for creating new users. Includes all the required
+    A form for creating a new user. Includes all the required
     fields, plus a repeated password.
     """
     password1 = forms.CharField(label='Password', widget=forms.PasswordInput)
@@ -18,7 +17,7 @@ class AdminUserCreationForm(forms.ModelForm):
         fields = ('email',)
 
     def clean_password2(self):
-        # assure that correct password was entered twice
+        # Assure that correct password was entered twice
         password1 = self.cleaned_data.get("password1")
         password2 = self.cleaned_data.get("password2")
         if password1 and password2 and password1 != password2:
@@ -36,8 +35,8 @@ class AdminUserCreationForm(forms.ModelForm):
 
 class AdminUserChangeForm(forms.ModelForm):
     """
-    A form for updating users. Includes all attributes of the user, but replaces the password field
-    with admin's disabled password hash display field.
+    A form for updating an existent user. Includes all attributes of the user, but replaces the password field
+    with a password hash display field.
     """
     password = ReadOnlyPasswordHashField(help_text=("Raw passwords are not stored, so there is no way to see "
                                                     "this user's password, but you can change the password "
@@ -45,7 +44,7 @@ class AdminUserChangeForm(forms.ModelForm):
 
     class Meta:
         model = User
-        # define fields which will be edited (not displayed)
+        # Define fields which will be edited (not displayed)
         fields = (
             'email', 'password', 'institute', 'bio',
             'is_staff', 'is_active', 'groups', 'user_permissions', 'last_login')
