+++ backend/core/user/viewsets.py
@@ -2,12 +2,11 @@ from core.user.serializers import UserSerializer
 from core.user.models import User
 from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated, AllowAny
-from rest_framework import filters
 
 
 class UserViewSet(viewsets.ModelViewSet):
     """
-    API endpoint that allows users to be viewed or edited.
+    API endpoint that allows users to be viewed or edited
     """
     queryset = User.objects.all().order_by('-date_joined')
     serializer_class = UserSerializer
